feat: set convert button regarding input field
All actions push a convert immediately except changes in the input
field. That's because every new character would do a conversion what
makes a lot of traffic. But to give a hint that a new conversion could
be useful the convert button will be enabled if the content is differ-
ent from the last conversion. All other actions trigger the conversion
and disable the convert button.

Signed-off-by: Conrad Zelck <git@simpel.cc>